Rearrange cubic interpolation
